epl_01/requirements.txt
flask
psycopg2-binary
python-dotenv
jsonify
flask_cors
